Remove 'Home Directory'(1.2).#214
