+++ jp.go.aist.rtm.rtcbuilder.nl1/plugin_ja.properties
@@ -37,9 +37,9 @@ perspectiveName = RTCBuilder
 preferecePagesName = RTCBuilder
 preferecePagesCodeGenerateName = \u30b3\u30fc\u30c9\u751f\u6210
 preferecePagesConfigurationName = Configuration
-preferecePagesPortsName = Port
-preferecePagesDocumentName = Document
-preferecePagesDataTypeName = Data Type
+preferecePagesPortsName = \u30dd\u30fc\u30c8
+preferecePagesDocumentName = \u30a2\u30af\u30c6\u30a3\u30d3\u30c6\u30a3, \u30c9\u30ad\u30e5\u30e1\u30f3\u30c8\u751f\u6210
+preferecePagesConfigurationName = \u30b3\u30f3\u30d5\u30a3\u30ae\u30e5\u30ec\u30fc\u30b7\u30e7\u30f3
 
 #Content Types
 rtcProfileName = RTC\u30d7\u30ed\u30d5\u30a1\u30a4\u30eb
