+++ jp.go.aist.rtm.rtcbuilder/plugin.properties
@@ -38,8 +38,8 @@ preferecePagesName = RTCBuilder
 preferecePagesCodeGenerateName = Code Generate
 preferecePagesConfigurationName = Configuration
 preferecePagesPortsName = Port
-preferecePagesDocumentName = Document
-preferecePagesDataTypeName = Data Type
+preferecePagesDocumentName = Activity, Document
+preferecePagesConfigurationName = Configuration
 
 #Content Types
 rtcProfileName = RTC Profile
