+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/nl/messages.properties
@@ -497,10 +497,6 @@ IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_TITLE=Project Generation
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P1=The specified project does not exist in Workspace.
 IRTCBMessageConstants.CONFIRM_PROJECT_GENERATE_P2=Are you sure you want to generate new project?
 
-IPreferenceMessageConstants.HOME_DIR=Home Directory
-IPreferenceMessageConstants.DATA_TYPE_BTN_ADD=Add
-IPreferenceMessageConstants.DATA_TYPE_BTN_DELETE=Delete
-
 IPreferenceMessageConstants.CODE_GEN_TITLE_BASIC=Basic
 IPreferenceMessageConstants.CODE_GEN_TITLE_DOCUMENT=Document
 IPreferenceMessageConstants.CODE_GEN_TITLE_CONFIG=Configuration
