+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/ui/editors/BasicEditorFormPage.java
@@ -117,7 +117,6 @@ public class BasicEditorFormPage extends AbstractEditorFormPage {
 	 * {@inheritDoc}
 	 */
 	protected void createFormContent(final IManagedForm managedForm) {
-		RTCUtil.setDefaultUserDir();
 		ScrolledForm form = super.createBase(managedForm, IMessageConstants.BASIC_SECTION);
 		FormToolkit toolkit = managedForm.getToolkit();
 		createModuleSection(toolkit, form);
